cleaned up unused values and extracted Slider for reusability
