+++ src/App.js
@@ -100,7 +100,7 @@ const App = props => {
     })
   }, []);
 
-if(!localStorage.getItem("token")){ // temp setting for testing purposes
+if(user.new_user){ // temp setting for testing purposes
   return(
   <>
   <UserContext.Provider value={{...user, setUser: setUser}}>
