+++ src/views/onboarding/Login.js
@@ -1,6 +1,6 @@
 // IMPORTS
 // react
-import React, { useState, useContext } from "react";
+import React, { useContext } from "react";
 // contexts
 import { UserContext } from "../../contexts/UserContext";
 // styled components
@@ -176,13 +176,6 @@ const HeaderHolder = styled.div`
   }
 `;
 
-const OnboardTxt = styled.p`
-  margin: auto;
-  font-size: 2rem;
-  line-height: 33px;
-  letter-spacing: 0.035em;
-  color: #ccc9ff;
-`;
 const FlexHolder = styled.div`
   display: flex;
   flex-direction: column;
@@ -215,10 +208,6 @@ const Button = styled.a`
   font-size:1.6rem;
   letter-spacing:0.1rem;
 }
-`;
-
-const LineTime = styled.hr`
-  width: 100%;
 `;
 
 const FBButton = styled.a`
@@ -266,10 +255,6 @@ const Cubes = styled.img`
 `;
 
 const Logo = styled.img``;
-
-const GoogleLogo = styled(Logo)`
-  border-radius: 50%;
-`;
 
 // EXPORT
 export default Login;
