+++ src/views/onboarding/steps/Height.js
@@ -155,9 +155,6 @@ const WeightContainer = styled.div`
     }
   }
 `;
-const DialStuff = styled.div`
-  flex: 0 0 auto;
-`;
 
 const NumberP = styled.p`
   margin-left: -1rem;
