+++ src/views/onboarding/steps/StepForm.js
@@ -10,8 +10,8 @@ import { axiosWithAuth } from "../../../helpers/axiosWithAuth";
 import styled from "styled-components";
 //images
 import waves from "../../../images/Onboarding/waves.svg";
-//react slider library
-import ReactSlider from "react-slider";
+//slider component
+import Slider from "../../../styles/global/Slider"
 //weight and height dials
 import Weight from "./Weight";
 import Height from "./Height";
@@ -20,7 +20,7 @@ import LoadingSpinner from "../../../components/atoms/spinner/spinner";
 
 const StepForm = props => {
   //context
-  const user = useContext(UserContext);
+  // const user = useContext(UserContext);
 
   //hooks
   const [questions, setQuestions] = useState([]);
@@ -40,7 +40,6 @@ const StepForm = props => {
 
   //set values for slider
   const handleChanges = value => {
-    console.log(value);
     setAnswer(value);
   };
 
@@ -71,13 +70,13 @@ const StepForm = props => {
       .then(res => console.log(res));
   };
 
-  //slider thumb and track
-  const Thumb = (props, state) => (
-    <StyledThumb {...props}>{state.valueNow}</StyledThumb>
-  );
-  const Track = (props, state) => {
-    return <StyledTrack {...props} index={state.index} value={7} />;
-  };
+  // //slider thumb and track
+  // const Thumb = (props, state) => (
+  //   <StyledThumb {...props}>{state.valueNow}</StyledThumb>
+  // );
+  // const Track = (props, state) => {
+  //   return <StyledTrack {...props} index={state.index} value={7} />;
+  // };
 
   //StepDot helper function for styled component
   const StepDotCount = currentStep => {
@@ -164,11 +163,9 @@ const StepForm = props => {
           )}
           {/* Slider Group , Questions 4 - 10*/}
           {currentStep >= 3 && (
-            <StyledSlider
+            <Slider
               defaultValue={3}
               max={7}
-              renderTrack={Track}
-              renderThumb={Thumb}
               onChange={handleChanges}
             />
           )}
@@ -194,33 +191,6 @@ const StepForm = props => {
 // STYLED COMPONENTS
 //Onboarding Reusable Styles
 // we abstract out reusable global styles later on -JC
-//slider style for react slider
-const StyledSlider = styled(ReactSlider)`
-  width: 100%;
-  height: 0.2rem;
-  margin: 8rem 0 13rem;
-`;
-
-//thumb style for react slider
-const StyledThumb = styled.div`
-  height: 2.5rem;
-  line-height: 25px;
-  width: 25px;
-  text-align: center;
-  background-color: #e05cb3;
-  color: #fff;
-  border-radius: 50%;
-  cursor: grab;
-  margin-top: -1rem;
-`;
-
-//track style for react slider
-const StyledTrack = styled.div`
-  top: 0;
-  bottom: 0;
-  background: ${props => (props.index === 1 ? "#ddd" : "#E05CB3")};
-  border-radius: 2rem;
-`;
 
 const OnBoardContainer = styled.div`
   display: flex;
